+++ src/main/kotlin/uk/gov/justice/digital/hmpps/managerecallsapi/controller/RecallsController.kt
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PatchMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
@@ -19,6 +20,7 @@ import uk.gov.justice.digital.hmpps.managerecallsapi.db.RecallRepository
 import uk.gov.justice.digital.hmpps.managerecallsapi.documents.dossier.DossierService
 import uk.gov.justice.digital.hmpps.managerecallsapi.documents.encodeToBase64String
 import uk.gov.justice.digital.hmpps.managerecallsapi.documents.lettertoprison.LetterToPrisonService
+import uk.gov.justice.digital.hmpps.managerecallsapi.documents.personName
 import uk.gov.justice.digital.hmpps.managerecallsapi.documents.recallnotification.RecallNotificationService
 import uk.gov.justice.digital.hmpps.managerecallsapi.domain.CourtId
 import uk.gov.justice.digital.hmpps.managerecallsapi.domain.NomsNumber
@@ -26,7 +28,11 @@ import uk.gov.justice.digital.hmpps.managerecallsapi.domain.PrisonId
 import uk.gov.justice.digital.hmpps.managerecallsapi.domain.RecallId
 import uk.gov.justice.digital.hmpps.managerecallsapi.domain.UserId
 import uk.gov.justice.digital.hmpps.managerecallsapi.domain.random
+import uk.gov.justice.digital.hmpps.managerecallsapi.service.CourtValidationService
+import uk.gov.justice.digital.hmpps.managerecallsapi.service.PrisonValidationService
 import uk.gov.justice.digital.hmpps.managerecallsapi.service.RecallDocumentService
+import uk.gov.justice.digital.hmpps.managerecallsapi.service.UpdateRecallService
+import uk.gov.justice.digital.hmpps.managerecallsapi.service.UserDetailsService
 import java.time.LocalDate
 import java.time.OffsetDateTime
 import java.util.UUID
@@ -39,7 +45,11 @@ class RecallsController(
   @Autowired private val recallNotificationService: RecallNotificationService,
   @Autowired private val recallDocumentService: RecallDocumentService,
   @Autowired private val dossierService: DossierService,
-  @Autowired private val letterToPrison: LetterToPrisonService
+  @Autowired private val letterToPrison: LetterToPrisonService,
+  @Autowired private val userDetailsService: UserDetailsService,
+  @Autowired private val updateRecallService: UpdateRecallService,
+  @Autowired private val prisonValidationService: PrisonValidationService,
+  @Autowired private val courtValidationService: CourtValidationService
 ) {
 
   @PostMapping("/recalls")
@@ -59,6 +69,22 @@ class RecallsController(
   @GetMapping("/recalls/{recallId}")
   fun getRecall(@PathVariable("recallId") recallId: RecallId): RecallResponse =
     recallRepository.getByRecallId(recallId).toResponse()
+
+  @PatchMapping("/recalls/{recallId}")
+  fun updateRecall(
+    @PathVariable("recallId") recallId: RecallId,
+    @RequestBody updateRecallRequest: UpdateRecallRequest
+  ): ResponseEntity<RecallResponse> =
+    if (prisonValidationService.isValidAndActive(updateRecallRequest.currentPrison) &&
+      prisonValidationService.isValid(updateRecallRequest.lastReleasePrison) &&
+      courtValidationService.isValid(updateRecallRequest.sentencingCourt)
+    ) {
+      ResponseEntity.ok(
+        updateRecallService.updateRecall(recallId, updateRecallRequest).toResponse()
+      )
+    } else {
+      ResponseEntity.badRequest().build()
+    }
 
   @GetMapping("/recalls/{recallId}/recallNotification/{userId}")
   fun getRecallNotification(
@@ -110,64 +136,65 @@ class RecallsController(
     @PathVariable("assignee") assignee: UserId
   ): RecallResponse =
     recallRepository.unassignRecall(recallId, assignee).toResponse()
+
+  fun Recall.toResponse() = RecallResponse(
+    recallId = this.recallId(),
+    nomsNumber = this.nomsNumber,
+    documents = this.documents.map { doc -> ApiRecallDocument(doc.id, doc.category, doc.fileName) },
+    recallLength = this.recallLength,
+    lastReleasePrison = this.lastReleasePrison,
+    lastReleaseDate = this.lastReleaseDate,
+    recallEmailReceivedDateTime = this.recallEmailReceivedDateTime,
+    localPoliceForce = this.localPoliceForce,
+    contraband = this.contraband,
+    contrabandDetail = this.contrabandDetail,
+    vulnerabilityDiversity = this.vulnerabilityDiversity,
+    vulnerabilityDiversityDetail = this.vulnerabilityDiversityDetail,
+    mappaLevel = this.mappaLevel,
+    sentenceDate = this.sentencingInfo?.sentenceDate,
+    licenceExpiryDate = this.sentencingInfo?.licenceExpiryDate,
+    sentenceExpiryDate = this.sentencingInfo?.sentenceExpiryDate,
+    sentencingCourt = this.sentencingInfo?.sentencingCourt,
+    indexOffence = this.sentencingInfo?.indexOffence,
+    conditionalReleaseDate = this.sentencingInfo?.conditionalReleaseDate,
+    sentenceLength = this.sentencingInfo?.sentenceLength?.let {
+      Api.SentenceLength(
+        it.sentenceYears,
+        it.sentenceMonths,
+        it.sentenceDays
+      )
+    },
+    bookingNumber = this.bookingNumber,
+    probationOfficerName = this.probationInfo?.probationOfficerName,
+    probationOfficerPhoneNumber = this.probationInfo?.probationOfficerPhoneNumber,
+    probationOfficerEmail = this.probationInfo?.probationOfficerEmail,
+    localDeliveryUnit = this.probationInfo?.localDeliveryUnit,
+    authorisingAssistantChiefOfficer = this.probationInfo?.authorisingAssistantChiefOfficer,
+    licenceConditionsBreached = this.licenceConditionsBreached,
+    reasonsForRecall = this.reasonsForRecall.toList(),
+    reasonsForRecallOtherDetail = this.reasonsForRecallOtherDetail,
+    agreeWithRecall = this.agreeWithRecall,
+    agreeWithRecallDetail = this.agreeWithRecallDetail,
+    currentPrison = this.currentPrison,
+    additionalLicenceConditions = this.additionalLicenceConditions,
+    additionalLicenceConditionsDetail = this.additionalLicenceConditionsDetail,
+    differentNomsNumber = this.differentNomsNumber,
+    differentNomsNumberDetail = this.differentNomsNumberDetail,
+    recallNotificationEmailSentDateTime = this.recallNotificationEmailSentDateTime,
+    dossierEmailSentDate = this.dossierEmailSentDate,
+    hasOtherPreviousConvictionMainName = this.hasOtherPreviousConvictionMainName,
+    hasDossierBeenChecked = this.hasDossierBeenChecked,
+    previousConvictionMainName = this.previousConvictionMainName,
+    assessedByUserId = this.assessedByUserId(),
+    bookedByUserId = this.bookedByUserId(),
+    dossierCreatedByUserId = this.dossierCreatedByUserId(),
+    dossierTargetDate = this.dossierTargetDate,
+    assignee = this.assignee(),
+    assigneeUserName = this.assignee()?.let { userDetailsService.get(it).personName() }
+  )
 }
 
 fun BookRecallRequest.toRecall() = Recall(::RecallId.random(), this.nomsNumber)
-
-fun Recall.toResponse() = RecallResponse(
-  recallId = this.recallId(),
-  nomsNumber = this.nomsNumber,
-  documents = this.documents.map { doc -> ApiRecallDocument(doc.id, doc.category, doc.fileName) },
-  recallLength = this.recallLength,
-  lastReleasePrison = this.lastReleasePrison,
-  lastReleaseDate = this.lastReleaseDate,
-  recallEmailReceivedDateTime = this.recallEmailReceivedDateTime,
-  localPoliceForce = this.localPoliceForce,
-  contraband = this.contraband,
-  contrabandDetail = this.contrabandDetail,
-  vulnerabilityDiversity = this.vulnerabilityDiversity,
-  vulnerabilityDiversityDetail = this.vulnerabilityDiversityDetail,
-  mappaLevel = this.mappaLevel,
-  sentenceDate = this.sentencingInfo?.sentenceDate,
-  licenceExpiryDate = this.sentencingInfo?.licenceExpiryDate,
-  sentenceExpiryDate = this.sentencingInfo?.sentenceExpiryDate,
-  sentencingCourt = this.sentencingInfo?.sentencingCourt,
-  indexOffence = this.sentencingInfo?.indexOffence,
-  conditionalReleaseDate = this.sentencingInfo?.conditionalReleaseDate,
-  sentenceLength = this.sentencingInfo?.sentenceLength?.let {
-    Api.SentenceLength(
-      it.sentenceYears,
-      it.sentenceMonths,
-      it.sentenceDays
-    )
-  },
-  bookingNumber = this.bookingNumber,
-  probationOfficerName = this.probationInfo?.probationOfficerName,
-  probationOfficerPhoneNumber = this.probationInfo?.probationOfficerPhoneNumber,
-  probationOfficerEmail = this.probationInfo?.probationOfficerEmail,
-  localDeliveryUnit = this.probationInfo?.localDeliveryUnit,
-  authorisingAssistantChiefOfficer = this.probationInfo?.authorisingAssistantChiefOfficer,
-  licenceConditionsBreached = this.licenceConditionsBreached,
-  reasonsForRecall = this.reasonsForRecall.toList(),
-  reasonsForRecallOtherDetail = this.reasonsForRecallOtherDetail,
-  agreeWithRecall = this.agreeWithRecall,
-  agreeWithRecallDetail = this.agreeWithRecallDetail,
-  currentPrison = this.currentPrison,
-  additionalLicenceConditions = this.additionalLicenceConditions,
-  additionalLicenceConditionsDetail = this.additionalLicenceConditionsDetail,
-  differentNomsNumber = this.differentNomsNumber,
-  differentNomsNumberDetail = this.differentNomsNumberDetail,
-  recallNotificationEmailSentDateTime = this.recallNotificationEmailSentDateTime,
-  dossierEmailSentDate = this.dossierEmailSentDate,
-  hasOtherPreviousConvictionMainName = this.hasOtherPreviousConvictionMainName,
-  hasDossierBeenChecked = this.hasDossierBeenChecked,
-  previousConvictionMainName = this.previousConvictionMainName,
-  assessedByUserId = this.assessedByUserId(),
-  bookedByUserId = this.bookedByUserId(),
-  dossierCreatedByUserId = this.dossierCreatedByUserId(),
-  dossierTargetDate = this.dossierTargetDate,
-  assignee = this.assignee()
-)
 
 data class BookRecallRequest(val nomsNumber: NomsNumber)
 
@@ -217,7 +244,8 @@ data class RecallResponse(
   val bookedByUserId: UserId? = null,
   val dossierCreatedByUserId: UserId? = null,
   val dossierTargetDate: LocalDate? = null,
-  val assignee: UserId? = null
+  val assignee: UserId? = null,
+  val assigneeUserName: String? = null
 ) {
   val recallAssessmentDueDateTime: OffsetDateTime? = recallEmailReceivedDateTime?.plusHours(24)
   val status: Status? = calculateStatus()
@@ -268,4 +296,91 @@ enum class Status {
   IN_ASSESSMENT,
   RECALL_NOTIFICATION_ISSUED,
   DOSSIER_ISSUED
+}
+
+data class UpdateRecallRequest(
+  val lastReleasePrison: PrisonId? = null,
+  val lastReleaseDate: LocalDate? = null,
+  val recallEmailReceivedDateTime: OffsetDateTime? = null,
+  val localPoliceForce: String? = null,
+  val contraband: Boolean? = null,
+  val contrabandDetail: String? = null,
+  val vulnerabilityDiversity: Boolean? = null,
+  val vulnerabilityDiversityDetail: String? = null,
+  val mappaLevel: MappaLevel? = null,
+  val sentenceDate: LocalDate? = null,
+  val licenceExpiryDate: LocalDate? = null,
+  val sentenceExpiryDate: LocalDate? = null,
+  val sentencingCourt: CourtId? = null,
+  val indexOffence: String? = null,
+  val conditionalReleaseDate: LocalDate? = null,
+  val sentenceLength: Api.SentenceLength? = null,
+  val bookingNumber: String? = null,
+  val probationOfficerName: String? = null,
+  val probationOfficerPhoneNumber: String? = null,
+  val probationOfficerEmail: String? = null,
+  val localDeliveryUnit: LocalDeliveryUnit? = null,
+  val authorisingAssistantChiefOfficer: String? = null,
+  val licenceConditionsBreached: String? = null,
+  val reasonsForRecall: Set<ReasonForRecall>? = null,
+  val reasonsForRecallOtherDetail: String? = null,
+  val agreeWithRecall: AgreeWithRecall? = null,
+  val agreeWithRecallDetail: String? = null,
+  val currentPrison: PrisonId? = null,
+  val additionalLicenceConditions: Boolean? = null,
+  val additionalLicenceConditionsDetail: String? = null,
+  val differentNomsNumber: Boolean? = null,
+  val differentNomsNumberDetail: String? = null,
+  val recallNotificationEmailSentDateTime: OffsetDateTime? = null,
+  val dossierEmailSentDate: LocalDate? = null,
+  val hasOtherPreviousConvictionMainName: Boolean? = null,
+  val hasDossierBeenChecked: Boolean? = null,
+  val previousConvictionMainName: String? = null,
+  val assessedByUserId: UserId? = null,
+  val bookedByUserId: UserId? = null,
+  val dossierCreatedByUserId: UserId? = null
+)
+
+enum class RecallLength {
+  FOURTEEN_DAYS,
+  TWENTY_EIGHT_DAYS
+}
+
+enum class MappaLevel(val label: String) {
+  NA("N/A"),
+  LEVEL_1("Level 1"),
+  LEVEL_2("Level 2"),
+  LEVEL_3("Level 3"),
+  NOT_KNOWN("Not Known"),
+  CONFIRMATION_REQUIRED("Confirmation Required")
+}
+
+enum class RecallType {
+  FIXED
+}
+
+@Suppress("unused")
+enum class ReasonForRecall {
+  BREACH_EXCLUSION_ZONE,
+  ELM_BREACH_EXCLUSION_ZONE,
+  ELM_BREACH_NON_CURFEW_CONDITION,
+  ELM_FURTHER_OFFENCE,
+  ELM_EQUIPMENT_TAMPER,
+  ELM_FAILURE_CHARGE_BATTERY,
+  FAILED_HOME_VISIT,
+  FAILED_KEEP_IN_TOUCH,
+  FAILED_RESIDE,
+  FAILED_WORK_AS_APPROVED,
+  POOR_BEHAVIOUR_ALCOHOL,
+  POOR_BEHAVIOUR_FURTHER_OFFENCE,
+  POOR_BEHAVIOUR_DRUGS,
+  POOR_BEHAVIOUR_NON_COMPLIANCE,
+  POOR_BEHAVIOUR_RELATIONSHIPS,
+  TRAVELLING_OUTSIDE_UK,
+  OTHER
+}
+
+enum class AgreeWithRecall {
+  YES,
+  NO_STOP
 }
